Imprimir ticket y vale tras emitir Vale
