+++ packages/contextos/src/tpv/venta/detalle/detalle.ts
@@ -1,11 +1,14 @@
 import { ejecutarListaProcesos, MetaModelo, publicar } from "@olula/lib/dominio.ts";
 
+import { getReportFactura } from "#/ventas/factura/infraestructura.ts";
 import { cambioClienteVentaVacio, metaCambioClienteVenta, metaVenta } from "#/ventas/venta/dominio.ts";
 import { ProcesarContexto } from "@olula/lib/diseño.js";
+import { imprimir_blob } from "@olula/lib/impresion.js";
 import { accionesListaEntidades, ListaEntidades, ProcesarListaEntidades } from "@olula/lib/ListaEntidades.js";
 import { CambioClienteFactura, LineaFactura, PagoVentaTpv, VentaTpv } from "../diseño.ts";
 import { ventaTpvVacia } from "../dominio.ts";
-import { getLineas, getPagos, getVenta, patchCambiarDescuento, patchFechaVenta, patchVenta, postEmitirVale, postLineaPorBarcode } from "../infraestructura.ts";
+import { getLineas, getPagos, getReportVale, getReportVenta, getVenta, patchCambiarDescuento, patchFechaVenta, patchVenta, postEmitirVale, postLineaPorBarcode } from "../infraestructura.ts";
+import { imprimirTicketOFactura, imprimirVale } from "./DetalleVentaTpv.tsx";
 
 
 export const cambioClienteFacturaVacio: CambioClienteFactura = cambioClienteVentaVacio;
@@ -303,11 +306,14 @@ export const emitirVale: ProcesarVentaTpv = async (contexto, payload) => {
     const venta = payload as VentaTpv
     await postEmitirVale(venta)
 
-    return pipeVentaTpv(contexto, [
+    const [ctx, eventos] = await pipeVentaTpv(contexto, [
         refrescarCabecera,
         refrescarPagos,
         abiertaOEmitidaContexto,
     ]);
+    await imprimirTicketOFactura(ctx.venta, ctx.pagos.lista);
+    await imprimirVale(ctx.venta.codigo);
+    return [ctx, eventos];
 }
 
 
@@ -373,6 +379,32 @@ export const cargarContexto: ProcesarVentaTpv = async (contexto, payload) => {
     } else {
         return getContextoVacio(contexto);
     }
+}
+
+export const imprimirTicketOFactura = async (venta: VentaTpv, pagos: PagoVentaTpv[]) => {
+    if (venta.cliente) {
+        const blob = await getReportFactura(venta.id);
+        const url = URL.createObjectURL(blob);
+        window.open(url, "_blank");
+    } else {
+        const blob = await getReportVenta(venta.id);
+        console.log('imprimiendo venta');
+        await imprimir_blob(blob);
+        console.log('imprimiendo venta impresa');
+        const pagos_vale = pagos.filter((p) => p.vale && p.saldoVale && p.saldoVale > 0);
+        for (const p of pagos_vale) {
+            console.log("Imprimiendo vale", p.vale);
+            await imprimirVale(p.vale!);
+        }
+    }
+}
+
+export const imprimirVale = async (idVale: string) => {
+    console.log("getReportVale llamado con", idVale);
+    const blob = await getReportVale(idVale);
+    console.log("blob recibido", blob?.size, blob?.type);
+    await imprimir_blob(blob);
+    console.log("imprimir_blob completado para vale");
 }
